+++ src/cjs/utils/parsers/svg/types.js
@@ -1,0 +1,16 @@
+"use strict";
+/**
+ * Defining some basic SVG command types.
+ *
+ * As these types do not define all very possible path data strings they should cover the most daily use-cases.
+ *
+ * For more insight see
+ *    https://www.w3.org/TR/SVG/paths.html
+ *
+ *
+ * @author  Ikaros Kappler
+ * @date    2022-12-21 (winter solstice night – yes the days & nights are dark and I have nothing else to do)
+ * @version 0.0.1-alpha
+ */
+Object.defineProperty(exports, "__esModule", { value: true });
+//# sourceMappingURL=types.js.map