+++ src/esm/utils/datastructures/Color.js
@@ -487,10 +487,12 @@ export class Color {
      * @return {Color} The color instance that's represented by the given string.
      */
     static parse(str) {
-        if (typeof str == "undefined")
+        if (typeof str == "undefined") {
             return null;
-        if ((str = str.trim().toLowerCase()).length == 0)
+        }
+        if ((str = str.trim().toLowerCase()).length == 0) {
             return null;
+        }
         if (str.startsWith("#"))
             return Color.makeHEX(str.substring(1, str.length));
         if (str.startsWith("rgb")) {